error correction and optimization

+++ src/main/java/ru/mail/polis/homework/simple/IntegerTask.java
@@ -49,7 +49,7 @@ public class IntegerTask {
      */
     public static long factorial(byte n) {
         long result = 1L;
-        for (int i = 2; i <= n; ++i)
+        for (byte i = 2; i <= n; ++i)
             result *= i;
         return result;
     }
